requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
openai==1.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
dnspython==2.4.2
httpx==0.25.2
requests==2.31.0 